# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DealerOwnerContext": {
        "properties": {
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "type": {
            "const": "dealer_organization",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "organization_id"
        ],
        "title": "DealerOwnerContext",
        "type": "object"
      },
...
      "ListingCreateRequest": {
        "properties": {
          "owner_context": {
            "discriminator": {
              "mapping": {
                "dealer_organization": "#/components/schemas/DealerOwnerContext",
                "personal": "#/components/schemas/PersonalOwnerContext"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PersonalOwnerContext"
              },
              {
                "$ref": "#/components/schemas/DealerOwnerContext"
              }
            ],
            "title": "Owner Context"
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          },
          "vehicle_type": {
            "enum": [
              "car",
              "bike"
            ],
            "title": "Vehicle Type",
            "type": "string"
          }
        },
        "required": [
          "owner_context",
          "vehicle_type"
        ],
        "title": "ListingCreateRequest",
        "type": "object"
      },
...
      "ListingPrivateResponse": {
        "properties": {
          "asking_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asking Price"
          },
          "bike_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BikeSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "car_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CarSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_user_id": {
            "format": "uuid",
            "title": "Created By User Id",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lifecycle_status": {
            "enum": [
              "draft",
              "removed"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "moderation_status": {
            "enum": [
              "not_started",
              "pending"
            ],
            "title": "Moderation Status",
            "type": "string"
          },
          "owner_organization_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Organization Id"
          },
          "owner_type": {
            "enum": [
              "user",
              "dealer_organization"
            ],
            "title": "Owner Type",
            "type": "string"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "publication_status": {
            "enum": [
              "private",
              "pending"
            ],
            "title": "Publication Status",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "submitted_listing_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted Listing Version"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          },
          "vehicle_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "vehicle_type": {
            "enum": [
              "car",
              "bike"
            ],
            "title": "Vehicle Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_type",
          "owner_user_id",
          "owner_organization_id",
          "created_by_user_id",
          "vehicle_type",
          "variant_id",
          "lifecycle_status",
          "publication_status",
          "moderation_status",
          "submitted_listing_version",
          "submitted_at",
          "title",
          "description",
          "asking_price",
          "currency",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ListingPrivateResponse",
        "type": "object"
      },
...
      "OwnershipVerificationStartRequest": {
        "additionalProperties": false,
        "properties": {
          "chassis": {
            "anyOf": [
              {
                "maxLength": 48,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chassis"
          },
          "expected_listing_version": {
            "minimum": 1.0,
            "title": "Expected Listing Version",
            "type": "integer"
          },
          "jurisdiction": {
            "maxLength": 16,
            "minLength": 2,
            "title": "Jurisdiction",
            "type": "string"
          },
          "ownership_basis": {
            "enum": [
              "registered_owner",
              "company_vehicle",
              "financed_or_leased",
              "inherited",
              "authorized_representative"
            ],
            "title": "Ownership Basis",
            "type": "string"
          },
          "registration": {
            "maxLength": 32,
            "minLength": 4,
            "title": "Registration",
            "type": "string"
          },
          "vin": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          }
        },
        "required": [
          "expected_listing_version",
          "jurisdiction",
          "registration",
          "ownership_basis"
        ],
        "title": "OwnershipVerificationStartRequest",
        "type": "object"
      },
      "OwnershipVerificationStartResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "canonical_vehicle_id": {
            "format": "uuid",
            "title": "Canonical Vehicle Id",
            "type": "string"
          },
          "capture_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Expires At"
          },
          "capture_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Url"
          },
          "listing_version": {
            "title": "Listing Version",
            "type": "integer"
          },
          "ownership_basis": {
            "enum": [
              "registered_owner",
              "company_vehicle",
              "financed_or_leased",
              "inherited",
              "authorized_representative"
            ],
            "title": "Ownership Basis",
            "type": "string"
          },
          "reused": {
            "title": "Reused",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "session_pending",
              "pending",
              "manual_review",
              "verified",
              "failed",
              "expired",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "attempt_id",
          "canonical_vehicle_id",
          "listing_version",
          "status",
          "ownership_basis",
          "reused",
          "capture_url",
          "capture_expires_at"
        ],
        "title": "OwnershipVerificationStartResponse",
        "type": "object"
      },
      "OwnershipVerificationStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "canonical_vehicle_id": {
            "format": "uuid",
            "title": "Canonical Vehicle Id",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "ownership_basis": {
            "enum": [
              "registered_owner",
              "company_vehicle",
              "financed_or_leased",
              "inherited",
              "authorized_representative"
            ],
            "title": "Ownership Basis",
            "type": "string"
          },
          "reused": {
            "title": "Reused",
            "type": "boolean"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "status": {
            "enum": [
              "session_pending",
              "pending",
              "manual_review",
              "verified",
              "failed",
              "expired",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "attempt_id",
          "canonical_vehicle_id",
          "status",
          "ownership_basis",
          "verified_at",
          "expires_at",
          "revoked_at",
          "failure_code",
          "reused",
          "updated_at"
        ],
        "title": "OwnershipVerificationStatusResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "correlation_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "WheelMatch API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/listings": {
      "post": {
        "operationId": "create_listing_api_v1_listings_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPrivateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Listing",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/v1/listings/{listing_id}/ownership-verification/start": {
      "post": {
        "operationId": "start_ownership_verification_api_v1_listings__listing_id__ownership_verification_start_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipVerificationStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipVerificationStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Ownership Verification",
        "tags": [
          "ownership verification"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/ownership-verification/status": {
      "get": {
        "operationId": "get_ownership_verification_status_api_v1_listings__listing_id__ownership_verification_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipVerificationStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ownership Verification Status",
        "tags": [
          "ownership verification"
        ]
      }
    },
...
    "/api/v1/me/identity-verification": {
      "get": {
        "operationId": "get_identity_verification_status_api_v1_me_identity_verification_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityVerificationStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Identity Verification Status",
        "tags": [
          "identity verification"
        ]
      }
    },
    "/api/v1/me/identity-verifications": {
      "post": {
        "operationId": "start_identity_verification_api_v1_me_identity_verifications_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityVerificationStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Identity Verification",
        "tags": [
          "identity verification"
        ]
      }
    },
...
  }
}